Extract shared HTTP primitives to @foundry-toolkit/shared/http
ApiRequestError, requestJson<T>, and buildCompendiumQuery were
near-identical copies in apps/dm-tool/electron/compendium/client.ts
and apps/player-portal/src/api/client.ts. Move them to a new
packages/shared/src/http.ts (./http subpath) so the wire contract has
one source of truth.

- dm-tool re-exports under the historical local names
  (CompendiumRequestError, buildSearchQuery) so existing callsites and
  `instanceof` checks keep narrowing correctly.
- player-portal re-exports ApiRequestError from its own client.ts so
  the 9 callsites that import it from there don't have to churn.
- Three dm-tool test assertions that pinned `name: 'CompendiumRequestError'`
  are updated to `'ApiRequestError'` to match the unified class name.

No runtime behaviour change. Bucket C-b1 of the cleanup audit;
useRemoteData<T> + the FeatPicker/ItemShopPicker/Crafting refactors
(b2) follow in their own PR after this one merges.
