+++ apps/dm-tool/electron/compendium/client.ts
@@ -2,11 +2,15 @@
 // no caching, no fallbacks. The factory in ./index.ts composes this with
 // the SQLite cache layer from @foundry-toolkit/db/pf2e.
 //
-// Shapes mirror apps/character-creator/src/api/client.ts; only the
-// compendium subset dm-tool actually calls is ported.
+// The transport primitives (`ApiRequestError`, `requestJson`,
+// `buildCompendiumQuery`) live in `@foundry-toolkit/shared/http` so
+// player-portal can share them. We re-export under the historical local
+// names (`CompendiumRequestError`, `buildSearchQuery`) so existing
+// callsites + `instanceof` checks keep working.
+
+import { buildCompendiumQuery, requestJson } from '@foundry-toolkit/shared/http';
 
 import type {
-  ApiError,
   CompendiumDocument,
   CompendiumMatch,
   CompendiumPack,
@@ -14,50 +18,15 @@ import type {
   CompendiumSource,
 } from './types.js';
 
-export class CompendiumRequestError extends Error {
-  readonly status: number;
-  readonly suggestion: string | undefined;
-
-  constructor(status: number, error: string, suggestion?: string) {
-    super(error);
-    this.name = 'CompendiumRequestError';
-    this.status = status;
-    this.suggestion = suggestion;
-  }
-}
+export {
+  ApiRequestError as CompendiumRequestError,
+  buildCompendiumQuery as buildSearchQuery,
+} from '@foundry-toolkit/shared/http';
 
 /** Strip a trailing slash if present. The mcp config path normaliser
  *  already does this on save, but callers may pass a raw value. */
 function normaliseBase(url: string): string {
   return url.replace(/\/+$/, '');
-}
-
-async function requestJson<T>(url: string): Promise<T> {
-  const res = await fetch(url, { headers: { Accept: 'application/json' } });
-  if (!res.ok) {
-    let body: ApiError = { error: `HTTP ${res.status.toString()}` };
-    try {
-      body = (await res.json()) as ApiError;
-    } catch {
-      // Response wasn't JSON — fall through with the status-only error.
-    }
-    throw new CompendiumRequestError(res.status, body.error, body.suggestion);
-  }
-  return (await res.json()) as T;
-}
-
-export function buildSearchQuery(opts: CompendiumSearchOptions): string {
-  const params = new URLSearchParams();
-  if (opts.q !== undefined && opts.q.length > 0) params.set('q', opts.q);
-  if (opts.packIds !== undefined && opts.packIds.length > 0) params.set('packId', opts.packIds.join(','));
-  if (opts.documentType !== undefined) params.set('documentType', opts.documentType);
-  if (opts.traits !== undefined && opts.traits.length > 0) params.set('traits', opts.traits.join(','));
-  if (opts.anyTraits !== undefined && opts.anyTraits.length > 0) params.set('anyTraits', opts.anyTraits.join(','));
-  if (opts.sources !== undefined && opts.sources.length > 0) params.set('sources', opts.sources.join(','));
-  if (opts.ancestrySlug !== undefined && opts.ancestrySlug.length > 0) params.set('ancestrySlug', opts.ancestrySlug);
-  if (opts.maxLevel !== undefined) params.set('maxLevel', opts.maxLevel.toString());
-  if (opts.limit !== undefined) params.set('limit', opts.limit.toString());
-  return params.toString();
 }
 
 export interface CompendiumHttpClient {
@@ -80,7 +49,7 @@ export function createCompendiumHttpClient(baseUrl: string): CompendiumHttpClien
 
   return {
     searchCompendium(opts) {
-      return requestJson<{ matches: CompendiumMatch[] }>(`${base}/api/compendium/search?${buildSearchQuery(opts)}`);
+      return requestJson<{ matches: CompendiumMatch[] }>(`${base}/api/compendium/search?${buildCompendiumQuery(opts)}`);
     },
 
     getCompendiumDocument(uuid) {
